+++ src/main/java/graphics/devMode/common/DevSubPage.java
@@ -1,0 +1,49 @@
+package graphics.devMode.common;
+
+import java.util.List;
+import java.awt.event.ActionListener;
+
+import org.jetbrains.annotations.Nullable;
+
+import graphicsEngine.colors.SimpleColorScheme;
+import graphicsEngine.panels.BorderProperties;
+import graphicsEngine.presets.panels.AbstractHeader;
+
+import graphics.common.CommonWindow;
+import graphics.devMode.common.header.Header;
+import graphics.devMode.common.header.HeaderButtonListener;
+
+/**
+ * TODO: finish this and add javadocs
+ */
+public abstract class DevSubPage extends DevPage {
+    private HeaderButtonListener headerButtonListener;
+
+    /**
+     * TODO: finish this javadoc
+     */
+    public DevSubPage(@Nullable CommonWindow window,
+                      @Nullable List<ActionListener> actionListenerList,
+                      @Nullable SimpleColorScheme colors) {
+        super(window, actionListenerList, colors);
+    }
+
+    //TODO: add javadoc
+    @Override
+    public boolean addParticularListener(@Nullable ActionListener listener) {
+        if (listener instanceof HeaderButtonListener) {
+            headerButtonListener = (HeaderButtonListener) listener;
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * TODO: finish this javadoc
+     */
+    @Override
+    public @Nullable AbstractHeader getHeader(@Nullable SimpleColorScheme headerColors,
+                                              @Nullable BorderProperties borderProperties) {
+        return new Header(headerColors, borderProperties, headerButtonListener);
+    }
+}
