+++ src/main/java/graphics/devMode/common/header/Header.java
@@ -1,4 +1,4 @@
-package graphics.devMode.page1.header;
+package graphics.devMode.common.header;
 
 import javax.swing.BoxLayout;
 
@@ -14,14 +14,14 @@ import graphics.common.header.Button_Menu;
 /**
  * TODO: finish this and add javadocs
  */
-public class CommonHeader extends SimpleHeader {
+public class Header extends SimpleHeader {
 
     /**
      * TODO: finish this javadoc
      */
-    public CommonHeader(@Nullable SimpleColorScheme colors,
-                        @Nullable BorderProperties borderProperties,
-                        @Nullable HeaderButtonListener buttonListener) {
+    public Header(@Nullable SimpleColorScheme colors,
+                  @Nullable BorderProperties borderProperties,
+                  @Nullable HeaderButtonListener buttonListener) {
         super(colors, borderProperties, buttonListener);
     }
 
@@ -36,8 +36,6 @@ public class CommonHeader extends SimpleHeader {
         setLayout(new BoxLayout(this, BoxLayout.X_AXIS));
         @Nullable HeaderButtonListener buttonListener = (HeaderButtonListener) simpleButtonListener;
         add(new Button_Menu(buttonListener));
-        add(new Button_SelectionPage(buttonListener));
-        add(new Button_Page1(buttonListener));
         add(new Button_Pause(buttonListener));
     }
 }
