+++ src/main/java/graphics/devMode/common/header/HeaderButtonListener.java
@@ -1,4 +1,4 @@
-package graphics.devMode.page1.header;
+package graphics.devMode.common.header;
 
 import java.awt.event.ActionEvent;
 
@@ -9,8 +9,6 @@ import staticData.StaticData;
 import graphics.common.CommonWindow;
 import graphics.common.header.SimpleHeaderButtonListener;
 import graphics.common.header.Button_Menu;
-import graphics.devMode.selectionPage.SelectionPage;
-import graphics.devMode.page1.Page1;
 
 /**
  * TODO: add javadocs
@@ -32,10 +30,6 @@ public class HeaderButtonListener extends SimpleHeaderButtonListener {
         switch (e.getActionCommand()) {
             case Button_Menu.ACTION_COMMAND ->
                     getCommonWindow().toggleOverlay();
-            case Button_SelectionPage.ACTION_COMMAND ->
-                    getCommonWindow().setActivePage(SelectionPage.getStaticPageKey());
-            case Button_Page1.ACTION_COMMAND ->
-                    getCommonWindow().setActivePage(Page1.getStaticPageKey());
             case Button_Pause.ACTION_COMMAND ->
                     StaticData.pause = !StaticData.pause;
             //add new header button actions here
